+++ tests/integration/instruments/test_urllib3.py
@@ -2,13 +2,12 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import logging
-import sys
 
-import certifi
 import httpretty
 import pytest
 import urllib3
 
+from scout_apm.compat import urllib3_cert_pool_manager
 from scout_apm.instruments.urllib3 import ensure_installed
 from tests.compat import mock
 from tests.tools import delete_attributes
@@ -16,11 +15,6 @@ from tests.tools import delete_attributes
 mock_not_attempted = mock.patch(
     "scout_apm.instruments.urllib3.have_patched_pool_urlopen", new=False
 )
-
-if sys.version_info >= (3, 0):
-    CERT_REQUIRED = "CERT_REQUIRED"
-else:
-    CERT_REQUIRED = b"CERT_REQUIRED"
 
 
 def test_ensure_installed_twice(caplog):
@@ -86,7 +80,7 @@ def test_request(tracked_request):
             httpretty.GET, "https://example.com/", body="Hello World!"
         )
 
-        http = urllib3.PoolManager(cert_reqs=CERT_REQUIRED, ca_certs=certifi.where())
+        http = urllib3_cert_pool_manager()
         response = http.request("GET", "https://example.com")
 
     assert response.status == 200
@@ -100,7 +94,7 @@ def test_request(tracked_request):
 def test_request_type_error(tracked_request):
     ensure_installed()
     with pytest.raises(TypeError):
-        http = urllib3.PoolManager(cert_reqs=CERT_REQUIRED, ca_certs=certifi.where())
+        http = urllib3_cert_pool_manager()
         connection = http.connection_from_host("example.com", scheme="https")
         connection.urlopen()
 
@@ -118,7 +112,7 @@ def test_request_no_absolute_url(caplog, tracked_request):
             httpretty.GET, "https://example.com/", body="Hello World!"
         )
 
-        http = urllib3.PoolManager(cert_reqs=CERT_REQUIRED, ca_certs=certifi.where())
+        http = urllib3_cert_pool_manager()
         response = http.request("GET", "https://example.com")
 
     assert response.status == 200